requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-JWT-Extended
Flask-Bcrypt
flask-restx
flask-cors
python-dotenv
requests
pydantic
psycopg2-binary
pytest